requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
idna==2.9
numpy==1.18.1
Pillow==7.0.0
PyExecJS==1.5.1
PyQt5==5.9.2
PyQt5-sip==12.7.1
requests==2.23.0
sip==4.19.8
six==1.14.0
soupsieve==2.0
urllib3==1.25.8
